requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiohttp-retry==2.8.3
aioify==0.3.2
aiojira==0.1.14
aiosignal==1.2.0
aiosonic==0.15.1
aiospread==1.0.0
aiowebsocket==1.0.0.dev2
alabaster==0.7.12
allure-combine==1.0.8
allure-pytest==2.12.0
allure-python-commons==2.12.0
ansicon==1.89.0
anyio==3.6.2
astroid==2.12.13
asttokens==2.0.8
async-class==0.5.0
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.1
attr==0.3.2
attrs==22.1.0
autopep8==1.5.5
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
beautifultable==1.1.0
black==22.12.0
blessed==1.19.1
bracex==2.3.post1
Brotli==1.0.9
browser-cookie3==0.16.2
bs4==0.0.1
cachetools==5.2.0
cattrs==22.2.0
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coloredlogs==15.0.1
coverage==7.0.1
cryptography==37.0.4
Cython==0.29.32
dataclass-dict-convert==1.7.1
dataclasses-json==0.5.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
demjson3==3.0.5
dill==0.3.6
docopt==0.6.2
docstring-to-markdown==0.11
docutils==0.17.1
docx2pdf==0.1.8
docx2txt==0.8
entrypoints==0.4
et-xmlfile==1.1.0
execnet==1.9.0
executing==0.10.0
fastapi==0.86.0
filelock==3.7.1
flake8-bugbear==22.9.23
flake8-noqa==1.2.9
frozenlist==1.3.1
ghp-import==2.1.0
goco==1.0.6
google-api-core==2.8.2
google-api-python-client==2.65.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.56.4
graphviz==0.20.1
griffe==0.24.1
gspread==5.7.2
h11==0.13.0
h2==4.1.0
har2case==0.3.1
har2py==0.1.2
hiredis==2.1.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.1
humanfriendly==10.0
hupper==1.11
hyperframe==6.0.1
idna==3.3
imagesize==1.4.1
iniconfig==1.1.1
inquirer==3.0.0
install==1.3.5
ipykernel==6.15.1
ipython==8.4.0
isort==5.10.1
jaraco.classes==3.2.3
jedi==0.17.2
jeepney==0.8.0
jieba==0.42.1
jinxed==1.2.0
jira==2.0.0
jmespath==0.9.5
json-tricks==3.16.1
jupyter-client==7.3.4
jupyter-core==4.11.1
keyring==23.9.3
lazy-object-proxy==1.7.1
libcst==0.4.9
livereload==2.6.3
loguru==0.4.1
lsprotocol==2022.0.0a9
luadata==1.0.5
lupa==1.14.1
lxml @ https://download.lfd.uci.edu/pythonlibs/archived/lxml-4.9.0-cp311-cp311-win_amd64.whl
lz4==4.0.2
Markdown==3.3.7
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mergedeep==1.3.4
mike==1.1.2
mkdocs==1.4.2
mkdocs-autorefs==0.4.1
mkdocs-callouts==1.8.1
mkdocs-literate-nav==0.5.0
mkdocs-material==8.5.11
mkdocs-material-extensions==1.1.1
mkdocstrings==0.19.0
mkdocstrings-python==0.8.2
module-wrapper==0.2.4
MonkeyType==22.2.0
more-itertools==8.14.0
multidict==6.0.2
mypy==0.991
mypy-extensions==0.4.3
mypy-mypyc==0.670
natsort==8.2.0
nest-asyncio==1.5.5
node-vm2==0.4.5
nodeenv==1.7.0
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.0
onecache==0.3.1
openai==0.25.0
openpyxl==3.0.9
orjson==3.8.3
outcome==1.2.0
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221124
parso==0.7.1
PasteDeploy==3.0.1
pathspec==0.9.0
pbkdf2==1.3
pbr==5.11.1
pickleshare==0.7.5
pipreqs==0.4.11
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==2.5.2
pluggy==0.13.1
poetry-version==0.1.5
pprofile==2.1.0
prompt-toolkit==3.0.30
protobuf==4.21.5
psutil==5.9.1
pure-eval==0.2.2
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycallgraph==1.0.1
pycares==4.3.0
pycodestyle==2.6.0
pycparser==2.21
pycryptodome==3.15.0
pycurl @ file:///C:/Users/enor.deng/Downloads/pycurl-7.45.1-cp311-cp311-win_amd64.whl
pycurl-requests==0.5.0
pydantic==1.8.2
pydocstyle==6.1.1
PyExecJS==1.5.1
pyflakes==2.2.0
pygls==1.0.0
Pygments==2.13.0
pygsheets==2.0.6
pyink==22.12.0
PyJWT==2.6.0
pylint==2.15.6
pylint-runner==0.6.0
pymdown-extensions==9.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyramid==2.0
pyreadline3==3.4.1
pyright==1.1.286
PySocks==1.7.1
pytest==7.2.0
pytest-asyncio==0.19.0
pytest-common-subject==1.0.6
pytest-cov==4.0.0
pytest-fixture-order==0.1.4
pytest-html==3.1.1
pytest-lambda==2.2.0
pytest-metadata==2.0.2
pytest-multithreading==1.0.4
pytest-multithreading-allure==1.0.8
pytest-parallel==0.1.1
pytest-sugar==0.9.6
pytest-xdist==3.1.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-editor==1.0.4
python-gitlab==3.8.0
python-jsonrpc-server==0.4.0
pytkdocs==0.16.1
pytomlpp @ file:///C:/Users/enor.deng/Downloads/pytomlpp-1.0.11-cp311-cp311-win_amd64.whl
pytz==2022.1
PyWavelets @ file:///C:/Users/enor.deng/Downloads/PyWavelets-1.3.0-cp311-cp311-win_amd64.whl
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyyaml_env_tag==0.1
pyzmq==23.2.1
readchar==4.0.3
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rope==1.6.0
rsa==4.9
SecretStorage==3.3.3
selenium==4.4.0
sentry-sdk==0.14.4
Shapely @ file:///C:/Users/enor.deng/Downloads/Shapely-1.8.2-cp311-cp311-win_amd64.whl
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sourcery-cli==1.0.1
Sphinx==5.3.0
sphinx-rtd-theme==1.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.4.0
starlette==0.20.4
stringcase==1.2.0
tblib==1.7.0
termcolor==1.1.0
tkcalendar==1.6.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.5.11
tornado==6.2
tornado-httpclient-session==0.2.5
tqdm==4.64.0
traitlets==5.3.0
translationstring==1.4
trio==0.21.0
trio-websocket==0.9.2
typed-ast==1.5.4
typeguard==2.13.3
types-psutil==5.9.5.5
types-pycurl==7.45.2.0
types-pytz==2022.6.0.1
types-setuptools==65.6.0.2
types-toml==0.10.8.1
types-typed-ast==1.5.8.3
typing-inspect==0.8.0
typing_extensions==4.3.0
tzdata==2022.2
ujson==5.6.0
uritemplate==4.1.1
urllib3==1.26.11
uvicorn==0.19.0
venusian==3.0.0
verspec==0.1.0
watchdog==2.1.9
wcmatch==8.4.1
wcwidth==0.2.5
WebOb==1.8.7
websocket-client==1.4.1
websockets==10.3
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.1.0
XlsxWriter==3.0.3
yapf==0.32.0
yarg==0.1.9
yarl==1.8.1
zipp==3.11.0
zope.deprecation==4.4.0
zope.interface==5.5.2
